+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
+- #1264 Makes testing-harness generic over runtime.
 - #1267 Fixes HTTP 500 in Ledger REST API `slots/<number>` endpoint.
 - #1261 Return Attestation and Challenge in ProofReceipt 
 - #1252 Add `Attestation` data in the `ProofReceiptContents::Attestation` variant.